requirements.txt
# FastAPI and server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP client for Gmail API
httpx>=0.24.0
requests>=2.31.0

# Celery client (for sending tasks)
celery==5.3.4
kombu==5.3.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Encryption (for token decryption)
cryptography>=41.0.0

# Environment variables
python-dotenv==1.0.0
